[Feat] Dvided Data class
